requirements.txt
streamlit
groq
PyPDF2
pdfplumber
tiktoken
faiss-cpu
python-dotenv 